Consolidate icons and images
and deleted toast.js
