--- src/components/files/icons.jsx
@@ -1,7 +1,0 @@
-import React from "react";
-
-// TODO: make svg icons and use SvgIcon as in https://v0.material-ui.com/#/components/svg-icon
-const GithubIcon = () => <img src='http://localhost:8000/static/icons/github-circle.png' height="48" width="48" alt=""/>;
-const SciGymIcon = () => <img src='http://localhost:8000/static/icons/scigym-logo.png' height="37" width="37" alt="" style={{padding: "5px"}}/>
-
-export {GithubIcon, SciGymIcon}
