+++ src/components/profile/RepositoryItem.jsx
@@ -8,8 +8,7 @@ import CardContent from '@material-ui/core/CardContent';
 import Typography from '@material-ui/core/Typography';
 import Button from '@material-ui/core/Button';
 
-import { GithubIcon } from '../files/icons';
-import { SciGymLogo } from '../files/images';
+import { SciGymLogo, GithubIcon } from '../files/images';
 
 const RepositoryItem = ({
   name,
